Extracted enums from vvplay_async.rs
