+++ src/vvplay_async_prefetch/args.rs
@@ -2,8 +2,13 @@ use clap::Parser;
 use std::ffi::OsString;
 use std::path:: PathBuf;
 
+
+use crate::vvplay_async_prefetch::enums::AbrType;
+use crate::vvplay_async_prefetch::enums::DecoderType;
+use crate::vvplay_async_prefetch::enums::ThroughputPredictionType;
+use crate::vvplay_async_prefetch::enums::ViewportPredictionType;
 /**
- * This file contains all the command line argument and enums defintion for vvplay_async_prefetch
+ * This file contains all the command line argumentfor vvplay_async_prefetch.rs
  */
 
 
@@ -77,39 +82,4 @@ pub struct Args {
     pub enable_fetcher_optimizations: bool,
     #[clap(long, default_value = "rgb(255,255,255)")]
     pub bg_color: OsString,
-}
-
-//Noop for operation that will not use a decoder
-#[derive(clap::ValueEnum, Clone, Copy)]
-pub enum DecoderType {
-    Noop,
-    Draco,
-    Tmc2rs,
-}
-
-#[derive(clap::ValueEnum, Clone, Copy)]
-pub enum AbrType {
-    Quetra,
-    QuetraMultiview,
-    Mckp,
-}
-
-#[derive(clap::ValueEnum, Clone, Copy)]
-pub enum ThroughputPredictionType {
-    /// Last throughput
-    Last,
-    /// Average of last 3 throughput
-    Avg,
-    /// ExponentialMovingAverage,
-    Ema,
-    /// Gradient Adaptive Exponential Moving Average
-    Gaema,
-    /// Low Pass Exponential Moving Average
-    Lpema,
-}
-
-#[derive(clap::ValueEnum, Clone, Copy)]
-pub enum ViewportPredictionType {
-    /// Last viewport
-    Last,
 }
