+++ src/vvplay_async_prefetch/mod.rs
@@ -1,5 +1,6 @@
 pub mod args;
 pub mod buffer_manager;
 pub mod camera_trace;
+pub mod enums;
 pub mod fetch_request;
 pub mod network_trace;
